+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - add test-cases for sanitize
 - convert digits to letter, add unit-tests
+- make work dir for running build
 
 ## 0.0.7 (2020-01-26)